Moved playerData to seperate file
